design: GNB, 메인 슬라이더 추가
