+++ client/src/App.css
@@ -1,0 +1,10 @@
+.main_slider .swiper-slide a { display: block; }
+.main_slider .swiper-slide img { max-width: 100%; height: 370px; object-fit: cover; }
+.main_slider .swiper-pagination { position: absolute; left: 71%; bottom: 20px; width: 50px; height: 22px; line-height: 22px; background-color: rgba(0, 0, 0, 0.15); border-radius: 12px; color: #fff; }
+
+.main_slider .swiper-button-next, .swiper-button-prev { display: block; margin-top: 0; top: 43%; width: 200px; height: 52px; transition: all 0.5s ease 0s; opacity: 0; }
+.main_slider .swiper-button-next:after { right: 100%; }
+.main_slider .swiper-button-prev:after { left: 100%; transform: rotate(180deg); }
+.main_slider .swiper-button-prev:after,.swiper-button-next:after { content:""; display: block; position: absolute; width: 52px; height: 52px; background: url("/images/button.svg") no-repeat; background-size: cover; }
+.main_slider .swiper-wrapper:hover ~ .swiper-button-prev,
+.main_slider .swiper-wrapper:hover ~ .swiper-button-next { opacity: 1; }
