+++ client/src/components/DetailProduct/ProductInfo.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Descriptions, Button } from 'antd';
+// import { Descriptions, Button } from 'antd';
 import { useDispatch } from 'react-redux';
 import { addCart } from '../../store/actions/user';
