+++ client/src/components/Landing/origin_CheckBox.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { Collapse, Checkbox } from 'antd';
+// import { Collapse, Checkbox } from 'antd';
 const { Panel } = Collapse;
 
 function CheckBox(props) {
