+++ client/src/components/Navbar/LeftMenu/index.js
@@ -1,0 +1,26 @@
+import React from 'react';
+// import { Menu } from 'antd';
+// const SubMenu = Menu.SubMenu;
+// const MenuItemGroup = Menu.ItemGroup;
+
+const Index = (props) => {
+  return (
+
+    // <Menu mode={props.mode}>
+    //   <Menu.Item key="mail">
+    //     <a href="/">Home</a>
+    //   </Menu.Item>
+    //   <SubMenu title={<span>Blogs</span>}>
+    //     <MenuItemGroup title="Item 1" />
+    //     <MenuItemGroup title="Item 2" />
+    //   </SubMenu>
+    // </Menu>
+
+
+    <>
+    </>
+
+  )
+}
+
+export default Index
