+++ client/src/pages/FooterPage/index.js
@@ -1,14 +1,14 @@
 import React from 'react'
-import {Icon} from 'antd';
+// import {Icon} from 'antd';
 
 const Index = () => {
     return (
         <div style={{
             height: '80px', display: 'flex',
             flexDirection: 'column', alignItems: 'center',
-            justifyContent: 'center', fontSize:'1rem'
+            justifyContent: 'center', fontSize: '1rem'
         }}>
-           <p> Happy Coding  <Icon type="smile" /></p>
+            <p> Happy Coding </p>
         </div>
     )
 }
