+++ client/src/pages/RegisterPage/index.js
@@ -6,11 +6,11 @@ import moment from "moment";
 import * as Yup from 'yup';
 import Auth from "../../hooks/auth";
 
-import {
-  Form,
-  Input,
-  Button,
-} from 'antd';
+// import {
+//   Form,
+//   Input,
+//   Button,
+// } from 'antd';
 
 const formItemLayout = {
   labelCol: {
