+++ client/src/reset.css
@@ -1,0 +1,103 @@
+@charset "utf-8"; * { box-sizing: border-box; }
+html,body { height: 100%; }
+body { font-size: 14px; line-height: 1.4; font-family: "Noto Sans KR", sans-serif; color: #333; }
+table,
+th,
+td,
+input,
+select,
+button,
+textarea,
+pre { font: inherit; color: inherit; }
+body,
+dl,
+dt,
+dd,
+ul,
+ol,
+li,
+h1,
+h2,
+h3,
+h4,
+h5,
+h6,
+pre,
+form,
+fieldset,
+input,
+select,
+button,
+textarea,
+p,
+blockquote,
+table,
+th,
+td { margin: 0; padding: 0; } /* 여백 초기화 */
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 { font-size: 100%; } /* 제목 태그 초기화 */
+ol,
+ul,
+li { list-style: none; } /* 블릿 기호 초기화 */
+img,
+fieldset { border: 0; }
+img { vertical-align: middle; }
+a { color: #222; text-decoration: none; }
+table { border-spacing: 0; border-collapse: collapse; border: 0; }
+th,
+td { word-break: break-all; word-wrap: break-word; }
+label { cursor: pointer; }
+input,
+button { border: 0; border-radius: 0; } /* 버튼 초기화 */
+textarea { resize: none; }
+input,
+button,
+select,
+textarea { vertical-align: middle; }
+address,
+caption,
+cite,
+code,
+dfn,
+em,
+var { font-style: normal; } /* 폰트 스타일 초기화 */
+hr { clear: both; display: none; }
+blockquote,
+q { quotes: none; }
+abbr,
+acronym { border: 0; }
+legend,
+caption { visibility: hidden; overflow: hidden; line-height: 0; width: 0; height: 0; margin: 0; padding: 0; text-indent: -99999px; font-size: 0; }
+article,
+aside,
+details,
+figcaption,
+figure,
+footer,
+header,
+hgroup,
+menu,
+nav,
+section,
+main { display: block; }
+img,
+video { width: 100%; } /* 반응형 초기화 */
+section { position: relative; margin: 0 auto; width: 1050px; }
+
+/* IR 효과 */
+.ir_pm { display: block; overflow: hidden; font-size: 0; line-height: 0; text-indent: -9999px; } /* 의미있는 이미지의 대체 텍스트를 제공하는 경우(Phark Method) */
+.ir_wa { display: block; overflow: hidden; position: relative; z-index: -1; width: 100%; height: 100%; } /* 의미있는 이미지의 대체 텍스트로 이미지가 없어도 대체 텍스트를 보여주고자 할 때(WA IR) */
+.ir_so { overflow: hidden; position: absolute; width: 0; height: 0; line-height: 0; text-indent: -9999px; } /* 대체 텍스트가 아닌 접근성을 위한 숨김 텍스트를 제공할 때 */
+
+/* clearfix */
+.clearfix { *zoom: 1; }
+.clearfix:before,
+.clearfix:after { display: block; content: ""; line-height: 0; }
+.clearfix::after { clear: both; }
+
+/* blind */
+.blind { position: absolute; visibility: hidden; overflow: hidden; height: 0; line-height: 0; width: 0; margin: 0; padding: 0; font: 0/0 Arial; }
