+++ package.json
@@ -33,6 +33,7 @@
     "react-redux": "^5.0.7",
     "redux-persist": "^6.0.0",
     "saslprep": "^1.0.3",
+    "sass": "^1.56.1",
     "supports-color": "^7.1.0"
   },
   "devDependencies": {
