+++ server/config/key.js
@@ -1,5 +1,5 @@
-if (process.env.NODE_ENV === 'production') {
-    module.exports = require('./prod');
-} else {
-    module.exports = require('./dev');
-}
+// if (process.env.NODE_ENV === 'production') {
+//     module.exports = require('./prod');
+// } else {
+//     module.exports = require('./dev');
+// }
